chore(SlotAttributePropagationController): rename controller files for consistency
