+++ 2nd-gen/packages/core/controllers/slot-attribute-propagation-controller/index.ts
