+++ 2nd-gen/packages/core/controllers/slot-attribute-propagation-controller/src/slot-attribute-propagation-controller.ts
